move mirror command to export
