+++ tests/export/bucket_test.py
@@ -1,11 +1,11 @@
-"""Unit tests for mirror command"""
+"""Unit tests for export bucket command"""
 import asyncio
 from unittest.mock import call, ANY
 
 import pytest
 from reduct import Client, Bucket, ReductError
 
-from .conftest import AsyncIter
+from tests.conftest import AsyncIter
 
 
 @pytest.fixture(name="dest_bucket")
@@ -18,7 +18,7 @@ def _make_dest_bucket(mocker) -> Bucket:
 
 @pytest.fixture(name="client")
 def _make_client(mocker, src_bucket, dest_bucket) -> Client:
-    kls = mocker.patch("reduct_cli.mirror.ReductClient")
+    kls = mocker.patch("reduct_cli.export.ReductClient")
     client = mocker.Mock(spec=Client)
     client.get_bucket.return_value = src_bucket
     client.create_bucket.return_value = dest_bucket
@@ -37,12 +37,12 @@ def walk_async_iterator(iterator: AsyncIter):
 
 
 @pytest.mark.usefixtures("set_alias")
-def test__mirror_ok(
+def test__export_bucket_ok(
     runner, conf, client, src_settings, src_bucket, dest_bucket, records
 ):  # pylint: disable=too-many-arguments
-    """Should mirror a bucket to another one"""
+    """Should export data from a bucket to another one"""
 
-    result = runner(f"-c {conf} mirror test/src_bucket test/dest_bucket")
+    result = runner(f"-c {conf} export bucket test/src_bucket test/dest_bucket")
     assert "Entry 'entry-1' (copied 1 records (6 B)" in result.output
     assert result.exit_code == 0
 
@@ -85,10 +85,10 @@ def test__mirror_ok(
 
 
 @pytest.mark.usefixtures("set_alias", "client")
-def test__mirror_ok_with_interval(runner, conf, src_bucket):
-    """Should mirror a bucket to another one with time interval"""
+def test__export_to_bucket_ok_with_interval(runner, conf, src_bucket):
+    """Should export data from a bucket to another one with time interval"""
     result = runner(
-        f"-c {conf} mirror --start 2022-01-02T00:00:01.100300+02:00 "
+        f"-c {conf} export bucket --start 2022-01-02T00:00:01.100300+02:00 "
         f"--stop 2022-02-01T00:00:00+02:00 "
         f"test/src_bucket test/dest_bucket"
     )
@@ -104,24 +104,24 @@ def test__mirror_ok_with_interval(runner, conf, src_bucket):
 def test__get_error(runner, conf, client):
     """Should print error if something got wrong"""
     client.get_bucket.side_effect = RuntimeError("Oops")
-    result = runner(f"-c {conf} mirror test/src_bucket test/dest_bucket")
+    result = runner(f"-c {conf} export bucket test/src_bucket test/dest_bucket")
     assert result.output == "[RuntimeError] Oops\nAborted!\n"
     assert result.exit_code == 1
 
 
 @pytest.mark.usefixtures("set_alias", "client", "src_bucket")
-def test__mirror_409(runner, conf, dest_bucket):
+def test__export_bucket_409(runner, conf, dest_bucket):
     """Should skip record if it already exists in destination bucket"""
     dest_bucket.write.side_effect = ReductError(409, "Conflict")
-    result = runner(f"-c {conf} mirror test/src_bucket test/dest_bucket")
+    result = runner(f"-c {conf} export bucket test/src_bucket test/dest_bucket")
     assert result.exit_code == 0
 
 
 @pytest.mark.usefixtures("set_alias", "client", "dest_bucket")
-def test__mirror_utc_timestamp(runner, conf, src_bucket):
+def test__export_bucket_utc_timestamp(runner, conf, src_bucket):
     """Should support Z designator for UTC timestamps"""
     result = runner(
-        f"-c {conf} mirror test/src_bucket test/dest_bucket "
+        f"-c {conf} export bucket test/src_bucket test/dest_bucket "
         f"--start 2022-01-02T00:00:01.100300Z --stop 2022-02-01T00:00:00Z"
     )
     assert result.exit_code == 0
@@ -131,20 +131,20 @@ def test__mirror_utc_timestamp(runner, conf, src_bucket):
 
 
 @pytest.mark.usefixtures("set_alias", "client", "dest_bucket")
-def test__mirror_specific_entry(runner, conf, src_bucket):
-    """Should mirror specific entry"""
+def test__export_bucket_specific_entry(runner, conf, src_bucket):
+    """Should export specific entry"""
     result = runner(
-        f"-c {conf} mirror test/src_bucket test/dest_bucket --entries=entry-2"
+        f"-c {conf} export bucket test/src_bucket test/dest_bucket --entries=entry-2"
     )
     assert result.exit_code == 0
     assert src_bucket.query.call_args_list == [call("entry-2", start=ANY, stop=ANY)]
 
 
 @pytest.mark.usefixtures("set_alias", "client", "dest_bucket")
-def test__mirror_multiple_specific_entry(runner, conf, src_bucket):
-    """Should mirror multiple specific entries"""
+def test__export_bucket_multiple_specific_entry(runner, conf, src_bucket):
+    """Should export bucket multiple specific entries"""
     result = runner(
-        f"-c {conf} mirror test/src_bucket test/dest_bucket --entries=entry-2,entry-1"
+        f"-c {conf} export bucket test/src_bucket test/dest_bucket --entries=entry-2,entry-1"
     )
     assert result.exit_code == 0
     assert src_bucket.query.call_args_list == [
